Add files via upload
Here is a much more complete codebase. I have removed the 9axis for now and tried to build all the needed functions feel free to edit them. We still need concrete turning values for the encoders i.e how many counts does it take to perform a turn. we also need a global ticker to order everything and once we have that down and have reliable movement we can implement the search algorithm